# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "agime-server",
    "description": "An AI agent",
    "contact": {
      "name": "AGIME",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/recipes/create": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "create_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRecipeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "412": {
            "description": "Precondition failed - Agent not available"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/recipes/decode": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "decode_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecodeRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe decoded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecodeRecipeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/recipes/delete": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "delete_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Recipe deleted successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Recipe not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/recipes/encode": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "encode_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncodeRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe encoded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncodeRecipeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/recipes/list": {
      "get": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "list_recipes",
        "responses": {
          "200": {
            "description": "Get recipe list successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRecipeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/recipes/parse": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "parse_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe parsed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseRecipeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid recipe format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/save": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "save_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Recipe saved to file successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveRecipeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/scan": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "scan_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe scanned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRecipeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/schedule": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "schedule_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe scheduled successfully"
          },
          "404": {
            "description": "Recipe not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/recipes/slash-command": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "set_recipe_slash_command",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSlashCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Slash command set successfully"
          },
          "404": {
            "description": "Recipe not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/sessions/{session_id}": {
      "get": {
        "tags": [
          "Session Management"
        ],
        "operationId": "get_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Session Management"
        ],
        "operationId": "delete_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session deleted successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorRequest": {
        "type": "object",
        "properties": {
          "contact": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "CreateRecipeRequest": {
        "type": "object",
        "required": [
          "session_id"
        ],
        "properties": {
          "author": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthorRequest"
              }
            ],
            "nullable": true
          },
          "session_id": {
            "type": "string"
          }
        }
      },
      "CreateRecipeResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "recipe": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Recipe"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "DecodeRecipeRequest": {
        "type": "object",
        "required": [
          "deeplink"
        ],
        "properties": {
          "deeplink": {
            "type": "string"
          }
        }
      },
      "DecodeRecipeResponse": {
        "type": "object",
        "required": [
          "recipe"
        ],
        "properties": {
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          }
        }
      },
      "DeleteRecipeRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
...
      "EncodeRecipeRequest": {
        "type": "object",
        "required": [
          "recipe"
        ],
        "properties": {
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          }
        }
      },
      "EncodeRecipeResponse": {
        "type": "object",
        "required": [
          "deeplink"
        ],
        "properties": {
          "deeplink": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ListRecipeResponse": {
        "type": "object",
        "required": [
          "manifests"
        ],
        "properties": {
          "manifests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeManifest"
            }
          }
        }
      },
...
      "ParseRecipeRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
      "ParseRecipeResponse": {
        "type": "object",
        "required": [
          "recipe"
        ],
        "properties": {
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          }
        }
      },
...
      "RecipeManifest": {
        "type": "object",
        "required": [
          "id",
          "recipe",
          "file_path",
          "last_modified"
        ],
        "properties": {
          "file_path": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "last_modified": {
            "type": "string"
          },
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          },
          "schedule_cron": {
            "type": "string",
            "nullable": true
          },
          "slash_command": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "SaveRecipeRequest": {
        "type": "object",
        "required": [
          "recipe"
        ],
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          }
        }
      },
      "SaveRecipeResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "ScanRecipeRequest": {
        "type": "object",
        "required": [
          "recipe"
        ],
        "properties": {
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          }
        }
      },
      "ScanRecipeResponse": {
        "type": "object",
        "required": [
          "has_security_warnings"
        ],
        "properties": {
          "has_security_warnings": {
            "type": "boolean"
          }
        }
      },
      "ScheduleRecipeRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "cron_schedule": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          }
        }
      },
...
      "SetSlashCommandRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "slash_command": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}